+++ Client/Models/Bank/BankAccount.cs
@@ -15,9 +15,9 @@ public class BankAccount
     public AppUser AppUser { get; set; }
 
     public string HolderName { get; set; }
-    public double SavingsBalance { get; set; }
-    public double CheckingsBalance { get; set; }
-    public double TransactionAmount { get; set; }
+    public decimal SavingsBalance { get; set; }
+    public decimal CheckingsBalance { get; set; }
+    public decimal TransactionAmount { get; set; }
 
 
 }
