+++ Client/Models/OrderFoodDTOs/OrderDetailDTO.cs
@@ -13,5 +13,5 @@ public class OrderDetailDTO
     public Food Food { get; set; }
 
     public int Count { get; set; }
-    public double PurchasePrice { get; set; }
+    public decimal PurchasePrice { get; set; }
 }
